@@ -97,36 +97,36 @@ class BaseDataDelete:
 class BaseDataDeleteTestCase(BaseDataDelete):
     def test_revoke_project_collection_user_acl(self):
         acl = "ils -A {}".format(self.project_collection_path)
-        ret_acl = subprocess.check_output(acl, shell=True)
+        ret_acl = subprocess.check_output(acl, shell=True, encoding="UTF-8")
         assert "{}#nlmumc".format(self.manager1) not in ret_acl
         assert "{}#nlmumc".format(self.manager2) not in ret_acl
 
-        assert "{}#nlmumc:read object".format("rods") in ret_acl
-        assert "{}#nlmumc:read object".format("service-disqover") in ret_acl
-        assert "{}#nlmumc:read object".format("service-pid") in ret_acl
+        assert "{}#nlmumc:read_object".format("rods") in ret_acl
+        assert "{}#nlmumc:read_object".format("service-disqover") in ret_acl
+        assert "{}#nlmumc:read_object".format("service-pid") in ret_acl
 
         # Check the ACL of a file in a sub-folder
         version_schema = formatters.format_schema_versioned_collection_path(self.project_id, self.collection_id, "1")
         acl_version_schema = "ils -A {}".format(version_schema)
-        ret_acl_version_schema = subprocess.check_output(acl_version_schema, shell=True)
+        ret_acl_version_schema = subprocess.check_output(acl_version_schema, shell=True, encoding="UTF-8")
         assert "{}#nlmumc".format(self.manager1) not in ret_acl_version_schema
         assert "{}#nlmumc".format(self.manager2) not in ret_acl_version_schema
 
-        assert "{}#nlmumc:read object".format("rods") in ret_acl_version_schema
-        assert "{}#nlmumc:read object".format("service-disqover") in ret_acl_version_schema
-        assert "{}#nlmumc:read object".format("service-pid") in ret_acl_version_schema
+        assert "{}#nlmumc:read_object".format("rods") in ret_acl_version_schema
+        assert "{}#nlmumc:read_object".format("service-disqover") in ret_acl_version_schema
+        assert "{}#nlmumc:read_object".format("service-pid") in ret_acl_version_schema
 
     def test_set_collection_deletion_metadata(self):
         metadata = "imeta ls -C {} {}".format(self.project_collection_path, DataDeletionAttribute.REASON.value)
-        ret_metadata = subprocess.check_output(metadata, shell=True)
+        ret_metadata = subprocess.check_output(metadata, shell=True, encoding="UTF-8")
         assert "value: {}".format(self.deletion_reason) in ret_metadata
 
         metadata = "imeta ls -C {} {}".format(self.project_collection_path, DataDeletionAttribute.DESCRIPTION.value)
-        ret_metadata = subprocess.check_output(metadata, shell=True)
+        ret_metadata = subprocess.check_output(metadata, shell=True, encoding="UTF-8")
         assert "value: {}".format(self.deletion_description) in ret_metadata
 
         metadata = "imeta ls -C {} {}".format(self.project_collection_path, DataDeletionAttribute.STATE.value)
-        ret_metadata = subprocess.check_output(metadata, shell=True)
+        ret_metadata = subprocess.check_output(metadata, shell=True, encoding="UTF-8")
         assert "value: {}".format(self.deletion_state) in ret_metadata
 
     def test_project_collection_metadata_removal_from_index(self):
