@@ -28,23 +28,23 @@ class BaseTestTapeArchiveMounted(BaseTestTapeArchive):
     @classmethod
     def add_archive_data_to_mounted_dropzone(cls):
         run_iquest = 'iquest "%s" "SELECT RESC_LOC WHERE RESC_NAME = \'{}\'"'.format(cls.ingest_resource)
-        remote_resource = subprocess.check_output(run_iquest, shell=True).strip()
+        remote_resource = subprocess.check_output(run_iquest, shell=True, encoding="UTF-8").strip()
         rule = "irule -r irods_rule_engine_plugin-python-instance -F /rules/utils/createFakeTapeFile.r '*dropzonePath=\"/mnt/ingest/zones/{}\"' '*remoteResource=\"{}\"'".format(
             cls.token, remote_resource
         )
         subprocess.check_call(rule, shell=True)
 
 
-class TestTapeArchiveS3Mounted(BaseTestTapeArchiveMounted):
+class TestTapeArchiveMountedS3(BaseTestTapeArchiveMounted):
     ingest_resource = "ires-hnas-umResource"
     destination_resource = "replRescUMCeph01"
 
 
-class TestTapeArchiveUMMounted(BaseTestTapeArchiveMounted):
+class TestTapeArchiveMountedUM(BaseTestTapeArchiveMounted):
     ingest_resource = "ires-hnas-umResource"
     destination_resource = "replRescUM01"
 
 
-class TestTapeArchiveAZMMounted(BaseTestTapeArchiveMounted):
+class TestTapeArchiveMountedAZM(BaseTestTapeArchiveMounted):
     ingest_resource = "ires-hnas-azmResource"
     destination_resource = "replRescAZM01"
